#241 - client - Move connections, executions and instances under “automation” dir
