+++ client/src/pages/automation/connections/oauth2/constants.ts
