+++ client/src/pages/automation/executions/Executions.tsx
@@ -1,5 +1,5 @@
-import LayoutContainer from '../../layouts/LayoutContainer/LayoutContainer';
-import PageHeader from '../../components/PageHeader/PageHeader';
+import LayoutContainer from '../../../layouts/LayoutContainer/LayoutContainer';
+import PageHeader from '../../../components/PageHeader/PageHeader';
 import React from 'react';
 
 export const Executions = () => {
