+++ client/src/pages/automation/instances/Instances.tsx
@@ -1,5 +1,5 @@
-import LayoutContainer from '../../layouts/LayoutContainer/LayoutContainer';
-import PageHeader from '../../components/PageHeader/PageHeader';
+import LayoutContainer from '../../../layouts/LayoutContainer/LayoutContainer';
+import PageHeader from '../../../components/PageHeader/PageHeader';
 import React from 'react';
 
 export const Instances = () => {
